requirements.txt
Flask
requests
python-dotenv
google-genai>=0.2.0
